backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kaggle==1.6.15
pandas==2.1.3
numpy==1.26.2
python-multipart==0.0.6
python-dotenv==1.0.0
